+++ Steady-Management.Domain/OrderDetail.cs
@@ -11,9 +11,9 @@ namespace Steady_Management.Domain
         private int orderId;
         private int productId;
         private int quantity;
-        private float unitPrice;
+        private decimal unitPrice;
 
-        public OrderDetail(int orderId, int productId, int quantity, float unitPrice)
+        public OrderDetail(int orderId, int productId, int quantity, decimal unitPrice)
         {
             this.orderId = orderId;
             this.productId = productId;
@@ -24,6 +24,6 @@ namespace Steady_Management.Domain
         public int OrderId { get => orderId; set => orderId = value; }
         public int ProductId { get => productId; set => productId = value; }
         public int Quantity { get => quantity; set => quantity = value; }
-        public float UnitPrice { get => unitPrice; set => unitPrice = value; }
+        public decimal UnitPrice { get => unitPrice; set => unitPrice = value; }
     }
 }
